Minor changes in preparation for EVN data

+++ vipcals/pipeline.py
@@ -33,6 +33,7 @@ import scripts.export_data as expo
 import scripts.phase_shift as shft
 
 from AIPSData import AIPSUVData, AIPSCat
+import Wizardry.AIPSData as wizard
 
 import functools
 print = functools.partial(print, flush=True)
@@ -260,35 +261,36 @@ def calibrate(filepath_list, filename_list, outpath_list, log_list, target_list,
           + 'been retrieved from\n' + good_url + '\n\nGC#1 created.\n')
         stats_df['need_gc'] = True
         
-    
+   
     if [1, 'AIPS FG'] not in uvdata.tables:
         try:
             retrieved_urls = tabl.load_fg_tables(uvdata)
-        except help.NoTablesError:
-            # If the pipeline finds no tables, stops here
-            print("No vlba.cal tables were found online. The pipeline will stop here.\n")
             for pipeline_log in log_list:
-                pipeline_log.write("\nNo vlba.cal tables were found online. The pipeline will stop here.\n")
-            return(1)
-        
-        for pipeline_log in log_list:
-            for good_url in retrieved_urls:
-                pipeline_log.write('Flag information was not available in the file, ' \
-                                    + 'it has been retrieved from ' + good_url + '\n')
-            pipeline_log.write('FG#1 created.\n')
+                for good_url in retrieved_urls:
+                    pipeline_log.write('Flag information was not available in the file, ' \
+                                        + 'it has been retrieved from ' + good_url + '\n')
+                pipeline_log.write('FG#1 created.\n')
 
-        # Move the flag file to the target folders
-        for path in outpath_list:
-            os.system('cp ../tmp/flags.vlba ' + path + '/TABLES/flags.vlba')
+            # Move the flag file to the target folders
+            for path in outpath_list:
+                os.system('cp ../tmp/flags.vlba ' + path + '/TABLES/flags.vlba')
 
-        # Clean the tmp directory
-        os.system('rm ../tmp/*.vlba')
-        
-        print('Flag information was not available in the file, ' \
-                        + 'it has been retrieved from\n' + good_url + '\n')
-        print('FG#1 created.\n')
-        stats_df['need_fg'] = True
-        stats_df['vlbacal_files'] = str(retrieved_urls)
+            # Clean the tmp directory
+            os.system('rm ../tmp/*.vlba')
+            
+            print('Flag information was not available in the file, ' \
+                            + 'it has been retrieved from\n' + good_url + '\n')
+            print('FG#1 created.\n')
+            stats_df['need_fg'] = True
+            stats_df['vlbacal_files'] = str(retrieved_urls)
+
+        except help.NoTablesError:
+            # If the pipeline finds no tables, gives a wrning but continues
+            print("No vlba.cal tables were found online. No initial flags will be applied.\n")
+            for pipeline_log in log_list:
+                pipeline_log.write("\nNo vlba.cal tables were found online. No initial flags will be applied.\n")
+            stats_df['need_fg'] = True
+            stats_df['vlbacal_files'] = None
 
     if missing_tables == True:
         t1 = time.time()
@@ -369,12 +371,16 @@ def calibrate(filepath_list, filename_list, outpath_list, log_list, target_list,
     seq = uvdata.seq
     
     t_avg = time.time()
-    ## If the time resolution is < 1s, average the dataset in time 
+    ## If the time resolution is < 2s, average the dataset in time 
     ## (unless other value is given)
-    try:
-        time_resol = float(uvdata.table('CQ', 1)[0]['time_avg'][0])
-    except TypeError: # Single IF datasets
-        time_resol = float(uvdata.table('CQ', 1)[0]['time_avg'])
+    if [1, 'AIPS CQ'] in uvdata.tables:
+        try:
+            time_resol = float(uvdata.table('CQ', 1)[0]['time_avg'][0])
+        except TypeError: # Single IF datasets
+            time_resol = float(uvdata.table('CQ', 1)[0]['time_avg'])
+    else:
+        wuvdata = wizard.AIPSUVData(uvdata.name, uvdata.klass, uvdata.disk, uvdata.seq)
+        time_resol = float(round(min([x.inttim for x  in wuvdata]), 2))
         
     if time_resol < time_aver:
         avgdata = AIPSUVData(aips_name[:9] + '_AT', uvdata.klass, disk_number, seq)
@@ -394,7 +400,7 @@ def calibrate(filepath_list, filename_list, outpath_list, log_list, target_list,
                             + '{:.2f}'.format(time_resol) \
                             + 's. It has been averaged to 2s.\n')
         print('\nThe time resolution was {:.2f}'.format(time_resol) \
-            + f's. It has been averaged to {time_aver}s.\n')
+            + f's. It has been averaged to {time_aver}s.')
         is_data_avg = True
         stats_df['time_avg'] = True
         stats_df['old_timesamp'] = time_resol
@@ -408,13 +414,24 @@ def calibrate(filepath_list, filename_list, outpath_list, log_list, target_list,
             
     ## If the channel bandwidth is smaller than 0.5 MHz, average the dataset 
     ## in frequency up to 0.5 MHz per channel (unless other value is given)
-    try:
-        ch_width = float(uvdata.table('CQ', 1)[0]['chan_bw'][0])
-        no_chan = int(uvdata.table('CQ', 1)[0]['no_chan'][0])
-    except TypeError: # Single IF datasets
-        ch_width = float(uvdata.table('CQ', 1)[0]['chan_bw'])
-        no_chan = int(uvdata.table('CQ', 1)[0]['no_chan'])
+    if [1, 'AIPS CQ'] in uvdata.tables:
+        try:
+            ch_width = float(uvdata.table('CQ', 1)[0]['chan_bw'][0])
+            no_chan = int(uvdata.table('CQ', 1)[0]['no_chan'][0])
+        except TypeError: # Single IF datasets
+            ch_width = float(uvdata.table('CQ', 1)[0]['chan_bw'])
+            no_chan = int(uvdata.table('CQ', 1)[0]['no_chan'])
         
+    else:
+        try:
+            ch_width = float(uvdata.table('FQ', 1)[0]['ch_width'][0])
+            total_width = float(uvdata.table('FQ', 1)[0]['total_bandwidth'][0])
+            no_chan = int(total_width/ch_width)
+        except TypeError: # Single IF datasets
+            ch_width = float(uvdata.table('FQ', 1)[0]['ch_width'])
+            total_width = float(uvdata.table('FQ', 1)[0]['total_bandwidth'])
+            no_chan = int(total_width/ch_width)
+
     if ch_width < freq_aver*1000:
         if is_data_avg == False:
             avgdata = AIPSUVData(aips_name[:9] + '_AF', uvdata.klass, \
@@ -753,13 +770,36 @@ def calibrate(filepath_list, filename_list, outpath_list, log_list, target_list,
     disp.write_box(log_list, 'Earth orientation parameters corrections')
     disp.print_box('Earth orientation parameters corrections')
 
-    eopc.eop_correct(uvdata)
+    with fits.open(filepath_list[0]) as hdul:
+        try:
+            if hdul[0].header['CORRELAT'].strip() == 'SFXC':
+                for pipeline_log in log_list:
+                    pipeline_log.write('\nEarth orientation parameter corrections cannot '\
+                                    + 'be applied for non-DiFX correlators.\n'\
+                                    + 'CL#3 will be copied from CL#2.\n')
+                    print('\nEarth orientation parameter corrections cannot be ' \
+                                    + 'applied for non-DiFX correlators.\n'\
+                                    + 'CL#3 will be copied from CL#2.\n')
+                    help.tacop(uvdata, 'CL', 2, 3)
+            else:
+                eopc.eop_correct(uvdata)
 
-    for pipeline_log in log_list:
-        pipeline_log.write('\nEarth orientation parameter corrections applied!\n'\
-                        + 'CL#3 created.\n')
-    print('\nEarth orientation parameter corrections applied!\nCL#3 created.\n')
-    os.system('rm -rf ../tmp/usno*')
+                for pipeline_log in log_list:
+                    pipeline_log.write('\nEarth orientation parameter corrections applied!\n'\
+                                    + 'CL#3 created.\n')
+                print('\nEarth orientation parameter corrections applied!\nCL#3 created.\n')
+                os.system('rm -rf ../tmp/usno*')
+
+        except KeyError:
+            eopc.eop_correct(uvdata)
+
+            for pipeline_log in log_list:
+                pipeline_log.write('\nEarth orientation parameter corrections applied!\n'\
+                                + 'CL#3 created.\n')
+            print('\nEarth orientation parameter corrections applied!\nCL#3 created.\n')
+            os.system('rm -rf ../tmp/usno*')
+
+
 
     # Counting visibilities
     expo.data_split(uvdata, target_list, cl_table=3, flagver=2)
@@ -1924,7 +1964,7 @@ def pipeline(input_dict):
 
                 # Define AIPS name
                 with fits.open(filepath_list_ID[0]) as hdul:
-                    aips_name = hdul[0].header['OBSERVER'] # + '_' + klass_1
+                    aips_name = hdul['UV_DATA'].header['OBSCODE'] # + '_' + klass_1
 
                 ## Check if the AIPS catalogue name is too long, and rename ##
                 # 12 is the maximum length for a file name in AIPS
@@ -1936,7 +1976,7 @@ def pipeline(input_dict):
                     aips_name_short = name[:size_name] + '_' + suffix
 
                 # Check if project directory already exists, if not, create one
-                project_dir = output_directory + '/' + hdul[0].header['OBSERVER']
+                project_dir = output_directory + '/' + hdul['UV_DATA'].header['OBSCODE']
                 if os.path.exists(project_dir) == False:
                     os.system('mkdir ' + project_dir)
 
@@ -2023,7 +2063,7 @@ def pipeline(input_dict):
 
         # Define AIPS name
         hdul = fits.open(filepath_list[0])
-        aips_name = hdul[0].header['OBSERVER'] # + '_' + klass_1
+        aips_name = hdul['UV_DATA'].header['OBSCODE'] # + '_' + klass_1
 
         ## Check if the AIPS catalogue name is too long, and rename ##
         aips_name_short = aips_name
@@ -2034,7 +2074,7 @@ def pipeline(input_dict):
             aips_name_short = name[:size_name] + '_' + suffix
 
         # Check if project directory already exists, if not, create one
-        project_dir = output_directory + '/' + hdul[0].header['OBSERVER']
+        project_dir = output_directory + '/' + hdul['UV_DATA'].header['OBSCODE']
         if os.path.exists(project_dir) == False:
             os.system('mkdir ' + project_dir)
 
@@ -2110,7 +2150,7 @@ def pipeline(input_dict):
 
         # Define AIPS name
         hdul = fits.open(filepath_list[0])
-        aips_name = hdul[0].header['OBSERVER'] # + '_' + klass_2
+        aips_name = hdul['UV_DATA'].header['OBSCODE'] # + '_' + klass_2
         
         ## Check if the AIPS catalogue name is too long, and rename ##
         aips_name_short = aips_name
@@ -2121,7 +2161,7 @@ def pipeline(input_dict):
             aips_name_short = name[:size_name] + '_' + suffix
 
         # Check if project directory already exists, if not, create one
-        project_dir = output_directory + '/' + hdul[0].header['OBSERVER']
+        project_dir = output_directory + '/' + hdul['UV_DATA'].header['OBSCODE']
         if os.path.exists(project_dir) == False:
             os.system('mkdir ' + project_dir)
 
@@ -2201,7 +2241,7 @@ def pipeline(input_dict):
 
         # Define AIPS name
         hdul = fits.open(filepath_list[0])
-        aips_name = hdul[0].header['OBSERVER'] 
+        aips_name = hdul['UV_DATA'].header['OBSCODE']
         
         ## Check if the AIPS catalogue name is too long, and rename ##
         aips_name_short = aips_name
@@ -2212,7 +2252,7 @@ def pipeline(input_dict):
             aips_name_short = name[:size_name] + '_' + suffix
 
         # Check if project directory already exists, if not, create one
-        project_dir = output_directory + '/' + hdul[0].header['OBSERVER']
+        project_dir = output_directory + '/' + hdul['UV_DATA'].header['OBSCODE']
         if os.path.exists(project_dir) == False:
             os.system('mkdir ' + project_dir)
